refactor(tree): import type from helper instead of index
