+++ packages/elements/src/tree/elements/tree.ts
@@ -9,8 +9,7 @@ import { List, valueFormatWarning } from '../../list/index.js';
 import { VERSION } from '../../version.js';
 import { defaultFilter } from '../helpers/filter.js';
 import { TreeRenderer } from '../helpers/renderer.js';
-import type { TreeData, TreeFilter } from '../helpers/types';
-import type { TreeDataItem } from '../index';
+import type { TreeData, TreeDataItem, TreeFilter } from '../helpers/types';
 import './tree-item.js';
 
 const EXPAND_TOGGLE_ATTR = 'expand-toggle';
