+++ packages/elements/src/tree/index.ts
@@ -1,6 +1,4 @@
-import type { TreeDataItem } from '@refinitiv-ui/utils/collection.js';
-
-import type { TreeData } from './helpers/types';
+import type { TreeData, TreeDataItem } from './helpers/types';
 
 export * from './elements/tree.js';
 export * from './elements/tree-item.js';
